Fixed some minor bugs in Frontendupload
git-svn-id: http://joomlacode.org/svn/sermon_speaker@23 5d6a53e0-5f49-0410-96cc-992822d07570

+++ SermonSpeaker3.4-Beta/com_sermonspeaker/admin/controllers/sermon.php
@@ -108,10 +108,10 @@ class SermonspeakerControllerSermon extends SermonspeakerController
 			if (!$row->load($cid[$i])){
 				$msg .= $row->getError();
 			} else {
-				if(file_exists(JPATH_ROOT.$row->sermon_path)){
+				if($row->sermon_path && file_exists(JPATH_ROOT.$row->sermon_path)){
 					$check = unlink(JPATH_ROOT.$row->sermon_path);
 					if (!$check){
-						$msg .= 'Error while deleting '.$file.'<br>';
+						$msg .= ' - Error while deleting '.$file;
 					}
 				}
 			}
